moved generation to utility class
